% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrplot.R
\name{floatMessagePart}
\alias{floatMessagePart}
\title{Prepares a float ascii string for the network message}
\usage{
floatMessagePart(floats)
}
\description{
Prepares a float ascii string for the network message
}
