% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrplot.R
\name{floatVectorMessage}
\alias{floatVectorMessage}
\title{Prepare a named vector of floats to be sent over the network
Puts length of the int message to the forefront of the message part}
\usage{
floatVectorMessage(name, floatVector)
}
\description{
Prepare a named vector of floats to be sent over the network
Puts length of the int message to the forefront of the message part
}
