% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrplot.R
\name{intVectorMessage}
\alias{intVectorMessage}
\title{Prepare a named vector of ints to be sent over the network
Puts length of the int message to the forefront of the message part}
\usage{
intVectorMessage(name, intVector)
}
\description{
Prepare a named vector of ints to be sent over the network
Puts length of the int message to the forefront of the message part
}
