% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrplot.R
\name{prepareHeader}
\alias{prepareHeader}
\title{Prepares the network header of a command}
\usage{
prepareHeader(
  plotTarget,
  command,
  floatVectorCount,
  intVectorCount,
  stringCount
)
}
\arguments{
\item{plotTarget}{Plotname to target the command to}

\item{command}{command type}

\item{floatVectorCount}{number of float vectors in the message}

\item{intVectorCount}{number of float vectors in the message}

\item{stringCount}{The number of strings in the message}
}
\description{
Prepares the network header of a command
}
