% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrplot.R
\name{selast}
\alias{selast}
\title{Calculates the explation percentages for each principal component in pr.out}
\usage{
selast(pr.out)
}
\arguments{
\item{host}{pr.out}
}
\description{
Calculates the explation percentages for each principal component in pr.out
}
