% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrplot.R
\name{sendMessage}
\alias{sendMessage}
\title{Sends the prepared ascii to the server
#'}
\usage{
sendMessage(message, host = "127.0.0.1", port = 8080)
}
\arguments{
\item{message}{The message to send}

\item{host}{The hostname of the server computer}

\item{port}{The tcp port to connect to}
}
\description{
Sends the prepared ascii to the server
#'
}
