% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrplot.R
\name{stringMessage}
\alias{stringMessage}
\title{Prepares a named string message for the network message}
\usage{
stringMessage(name, msgString)
}
\description{
Prepares a named string message for the network message
}
