% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrplot.R
\name{stringMessagePart}
\alias{stringMessagePart}
\title{Prepares string to be sent over the network
Puts string length to the forefront of the message part}
\usage{
stringMessagePart(msgString)
}
\description{
Prepares string to be sent over the network
Puts string length to the forefront of the message part
}
