% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrplot.R
\name{vrclear}
\alias{vrclear}
\title{Clears a specified vr plot from the server}
\usage{
vrclear(plotTarget, host = "localhost")
}
\arguments{
\item{plotTarget}{Plotname to target the command to}

\item{host}{The hostname of the plotting server computer}
}
\description{
Prepares a clear network message to be sent to the vr plotter server
and sends it.
}
