% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrplot.R
\name{vrlabels}
\alias{vrlabels}
\title{Set labels of a vr plot}
\usage{
vrlabels(plotTarget, xTitle, yTitle, zTitle, host = "localhost")
}
\arguments{
\item{plotTarget}{Plotname to target the command to}

\item{xTitle}{X axis title}

\item{yTitle}{Y axis title}

\item{zTitle}{Z axis title}

\item{host}{The hostname of the plotting server computer}
}
\description{
Prepares an axis title network message and sends it to the vr plotter server
}
