% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrplot.R
\name{vrplot}
\alias{vrplot}
\title{Set labels of a vr plot}
\usage{
vrplot(
  plotTarget,
  x,
  y,
  z,
  colorIndices,
  type = "sphere",
  size = 1,
  host = "localhost"
)
}
\arguments{
\item{plotTarget}{Plotname to target the command to}

\item{host}{The hostname of the plotting server computer}

\item{xTitle}{X axis title}

\item{yTitle}{Y axis title}

\item{zTitle}{Z axis title}
}
\description{
Prepares an axis title network message sends it to the vr plotter server
}
