% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrplot.R
\name{vrshowpca}
\alias{vrshowpca}
\title{Sends pca results from princomp command to vr plotting server}
\usage{
vrshowpca(plotTarget, title, pr.out, colors, type = "spheres")
}
\arguments{
\item{plotTarget}{Plotname to target the command to}

\item{title}{The main title of the plot}

\item{pr.out}{Result object from princomp}

\item{type}{Plot object type spheres or cubes}

\item{host}{The hostname of the plotting server computer}

\item{color}{A vector of color indices to use for plotting}
}
\description{
Sends pca results from princomp command to vr plotting server
}
