% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrplot.R
\name{vrtitle}
\alias{vrtitle}
\title{Set labels of a vr plot}
\usage{
vrtitle(plotTarget, title, host = "localhost")
}
\arguments{
\item{plotTarget}{Plotname to target the command to}
}
\description{
Prepares a main title network message and sends it to the vr plotter server
}
